refactor performers package
