+++ core/src/main/java/net/sourceforge/html5val/ValidateFormAttributeModelDelayedProcessor.java
@@ -1,14 +1,8 @@
 package net.sourceforge.html5val;
 
-import java.util.List;
-
 import org.thymeleaf.context.ITemplateContext;
 import org.thymeleaf.engine.AttributeName;
-import org.thymeleaf.exceptions.TemplateProcessingException;
 import org.thymeleaf.model.IModel;
-import org.thymeleaf.model.IModelFactory;
-import org.thymeleaf.model.IOpenElementTag;
-import org.thymeleaf.model.IProcessableElementTag;
 import org.thymeleaf.processor.element.AbstractAttributeModelProcessor;
 import org.thymeleaf.processor.element.IElementModelStructureHandler;
 import org.thymeleaf.templatemode.TemplateMode;
